# openapi.yaml
openapi: "3.1.0"
info:
  title: Genpic Platform API
  version: "0.1.0"
  description: |
    Genpic image generation platform HTTP API.

    **Primary generation**: `POST /api/generate` — the server routes `model` to
    OpenAI, Gemini, Wan, or the optional 祥云 (`xiangyun/…`) meta-provider. Requests
    normally include `base_url` and `api_key` (same as the embedded web form); each
    adapter prefers those values and falls back to `config.yaml` when they are
    omitted. **祥云** may omit them to use only server defaults; if provided, the
    same credentials apply to each backend in the configured try order.

    **Session authentication** (when MySQL is configured):
    - Cookie `genpic_session` — HTTP-only session token set by `POST /api/auth/register`
      or `POST /api/auth/login`. Send requests with `credentials: include` from browsers.
    - Authenticated requests resolve the signed-in user’s id for job ownership.

    **Job owner headers** (optional; still supported):
    - `X-Genpic-User-Id` — opaque user id (legacy / external auth integration).
    - `X-Genpic-Session` — stable per-browser id for anonymous history; required for
      migrating anonymous jobs on first login when sent alongside login/register.
    Jobs are listed and fetched only when the same cookie/session/user scope matches.

    Put the service behind your own gateway or network controls for production where needed.
...
  contact:
    name: Genpic Engineering
  license:
    name: Proprietary
servers:
...
paths:
...
  /api/generate:
    post:
      operationId: apiGenerate
      tags: [generate]
      summary: Generate images (async job + poll, or sync without job store)
      description: |
        **cmd/genpic** (job store always enabled): enqueues a generation job and returns
        **202 Accepted** with a job object (`id`, `status`, …). Poll `GET /jobs/{id}`
        until `status` is `succeeded` or `failed`; on success, `data` holds images
        (`url` and/or `b64_json`, `mime_type`, `revised_prompt`).

        **mvp_lite** (no job store): waits for the upstream and returns **200** with
        `created`, `data`, and `x_provider` (same normalized shape as the synchronous path).

        **Model routing** (by `model` id): OpenAI Images, Gemini generateContent, or Wan
        multimodal — see provider docs in-repo.

        Send `X-Genpic-Session` and/or `X-Genpic-User-Id` so jobs are scoped for listing.

        When **object storage** and MySQL are configured, logged-in users upload reference
        images and generation outputs to OSS; **400** with `code: storage_quota_exceeded`
        is returned if the request would exceed the user's quota.
      parameters:
        - $ref: "#/components/parameters/GenpicUserId"
        - $ref: "#/components/parameters/GenpicSession"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiGenerateRequest"
            examples:
              gpt_image:
                summary: GPT Image 2 — basic text-to-image
                value:
                  base_url: "https://api.openai.com/v1"
                  api_key: "sk-…"
                  model: gpt-image-2
                  prompt: "A serene mountain lake at golden hour, photorealistic"
                  n: 1
                  size: "1024x1024"
                  quality: high
                  response_format: url
              gemini_flash:
                summary: Gemini 2.5 Flash — portrait
                value:
                  base_url: "https://generativelanguage.googleapis.com/v1beta"
                  api_key: "…"
                  model: gemini-2.5-flash-image
                  prompt: "Studio portrait of a fox wearing a graduation cap"
                  n: 1
                  aspect_ratio: "1:1"
                  image_size: "512"
                  response_format: b64_json
      responses:
        "200":
          description: Images returned synchronously (e.g. MVP Lite)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageGenerationResponse"
        "202":
          description: Job accepted (cmd/genpic); poll GET /jobs/{id}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationJob"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "502":
          $ref: "#/components/responses/UpstreamError"

  # ── Chat completions (M2) ──────────────────────
  /v1/chat/completions:
    post:
      operationId: chatCompletions
      tags: [chat]
      summary: Generate images via OpenAI-compatible chat completions
      description: |
        Translates a standard OpenAI chat completions request into an image generation
        call. Intended for Cherry Studio, AI as Workspace, Lobe Chat, and any
        OpenAI-SDK client that only speaks the chat completions surface.

        **Routing**: the `model` field must identify an image-generation model
        (e.g. `gemini/gemini-2.5-flash-image`, `openai/gpt-image-2`, `wan/wan2.7-image`).

        **Credentials**:
        - `Authorization: Bearer <token>` — forwarded as the upstream `api_key`.
        - `X-Base-Url` header — overrides the server-configured upstream base URL.
        - If the server already has credentials in `config.yaml`, the Bearer token
          and header are optional (Mode A: server-side key).

        **Reference images**: include `type:"image_url"` parts with `data:` URIs in
        the last user message content to supply reference images.

        **Limitations**: `stream:true` is not supported; the response is always
        synchronous. Generation can take 30–180 s depending on the model.
      security: []
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
          description: |
            `Bearer <upstream_api_key>` — forwarded to the upstream provider.
            Optional when the server has credentials configured in config.yaml.
        - name: X-Base-Url
          in: header
          required: false
          schema:
            type: string
          description: Override the upstream base URL for this request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionsRequest"
      responses:
        "200":
          description: Synchronous chat completions response with generated image(s)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "502":
          $ref: "#/components/responses/UpstreamError"

  # ── Jobs ───────────────────────────────────────
  /jobs/{job_id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get generation job status and result
      description: |
        Poll after `202 Accepted` from `POST /api/generate`. When `status == succeeded`,
        `data` contains generated images.

        Returns **404** when the job exists but does not belong to the caller's
        scope (session cookie, `X-Genpic-User-Id`, or `X-Genpic-Session`).
      parameters:
        - $ref: "#/components/parameters/GenpicUserId"
        - $ref: "#/components/parameters/GenpicSession"
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: Job `id` from the `202` response body
      responses:
        "200":
          description: Job record (any status)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationJob"
        "404":
          $ref: "#/components/responses/NotFound"
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List recent generation jobs (newest first)
      description: |
        Jobs are filtered by caller scope: authenticated users (via `genpic_session`
        cookie) see their jobs; with `X-Genpic-User-Id` only that user's jobs;
        with `X-Genpic-Session` only anonymous jobs for that session; when neither
        applies as owner scope, only legacy jobs with no stored owner are listed.
      parameters:
        - $ref: "#/components/parameters/GenpicUserId"
        - $ref: "#/components/parameters/GenpicSession"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
          description: Pagination cursor (opaque, from previous response)
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/JobStatus"
      responses:
        "200":
          description: Paginated job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobList"

  # ── Auth & settings ────────────────────────────
  /api/auth/register:
    post:
      operationId: authRegister
      tags: [auth]
      summary: Register a new account (sets session cookie)
      security: []
      parameters:
        - $ref: "#/components/parameters/GenpicSession"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterRequest"
      responses:
        "201":
          description: User created; genpic_session cookie set
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          description: Auth disabled (no database configured)
  /api/auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Log in (sets session cookie)
      security: []
      parameters:
        - $ref: "#/components/parameters/GenpicSession"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Login succeeded; genpic_session cookie set
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: Auth disabled (no database configured)
...
  /api/auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Current authenticated user
      parameters:
        - $ref: "#/components/parameters/GenpicSessionCookie"
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: Auth disabled
  /api/user/settings:
    get:
      operationId: getUserSettings
      tags: [auth]
      summary: Read privacy settings for the signed-in user
      parameters:
        - $ref: "#/components/parameters/GenpicSessionCookie"
      responses:
        "200":
          description: Settings object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: Auth disabled
    put:
      operationId: updateUserSettings
      tags: [auth]
      summary: Update privacy settings (partial body allowed)
      parameters:
        - $ref: "#/components/parameters/GenpicSessionCookie"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSettingsPatch"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: Auth disabled
  /api/jobs/{job_id}/visibility:
    put:
      operationId: setJobVisibility
      tags: [community]
      summary: Set job visibility for community feed (author only)
      parameters:
        - $ref: "#/components/parameters/GenpicSessionCookie"
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: Job id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobVisibilityBody"
      responses:
        "200":
          description: Updated job (includes prompt and params for owner)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationJob"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List templates for a model
      description: |
        Returns public templates for the model, plus the caller's private templates when
        authenticated via `genpic_session`. Anonymous callers see only public rows.
        Templates are stored with wire-style `primary_model` (no `openai/`… prefix); pass the
        same `model` value as the SPA model selector (catalog id) and the server matches both
        forms.
      parameters:
        - $ref: "#/components/parameters/GenpicSessionCookie"
        - name: model
          in: query
          required: true
          schema:
            type: string
          description: SPA model selector value (catalog id like `openai/gpt-image-2` or wire id); matched against stored wire-style `primary_model` and the original string
      responses:
        "200":
          description: Template list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Save a succeeded job as a template
      description: |
        Owner-only. Optional `reference_images` should be supplied from the client when
        reference inputs were not persisted on the job record. Administrators may set
        `visibility` to `public` to publish a shared preset (requires `auth.admin_emails`
        / `GENPIC_ADMIN_EMAILS` configuration server-side).
      parameters:
        - $ref: "#/components/parameters/GenpicSessionCookie"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateRequest"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationTemplate"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: A template for this job id already exists (`code` template_exists_for_job)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "503":
          description: Templates disabled (no database) or job store unavailable
  /api/templates/{id}:
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template
      description: |
        Deletes a private template owned by the caller. Administrators may also delete
        any public template.
      parameters:
        - $ref: "#/components/parameters/GenpicSessionCookie"
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [status, id]
                properties:
                  status:
                    type: string
                    enum: [deleted]
                  id:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: Templates disabled
  /api/admin/users/reset-password:
    post:
      operationId: adminResetUserPassword
      tags: [admin]
      summary: Set a user's password (admin only)
      description: |
        Requires a signed-in user whose email is listed in `auth.admin_emails` / `GENPIC_ADMIN_EMAILS`.
        Clears all active sessions for the target user so they must log in again.
      parameters:
        - $ref: "#/components/parameters/GenpicSessionCookie"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminResetPasswordRequest"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPasswordResetResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          description: Auth disabled
...
components:
  parameters:
    GenpicUserId:
      name: X-Genpic-User-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
      description: |
        Opaque user identifier when the caller is authenticated outside Genpic.
        Jobs created with this header are listed only when the same value is sent.
    GenpicSession:
      name: X-Genpic-Session
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
      description: |
        Stable random id per browser or device (for example a UUID in localStorage).
        Used for anonymous generation history when `X-Genpic-User-Id` is not set.
...
  schemas:
...
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed]
...
  responses:
    BadRequest:
      description: Malformed request (missing fields, invalid JSON)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
...